+++ src/base/orientation/Orientation.ts
@@ -1,0 +1,16 @@
+export enum Orientation {
+  /**
+           ┌─┐
+       ┌─┐ │ │
+       │ │ │ │
+      ─┴─┴─┴─┴─ 
+      */
+  Horizontal = 'HORIZONTAL',
+  /**
+      ├─────┐
+      ├─────┘
+      ├────────┐
+      ├────────┘
+      */
+  Vertical = 'VERTICAL',
+}
