+++ src/charts/ScenarioChart.ts
@@ -3,8 +3,9 @@ import { ColumnAxis } from '../base/Axis';
 import { Box } from '../base/Box';
 import { Scenario, extractParallelValues } from '../base/Data';
 import { ParallelDataType } from '../base/DataTypes';
-import { Orientation, OrientationBox } from '../base/OrientationBox';
-import { OrientationText } from '../base/OrientationText';
+import { Orientation } from '../base/orientation/Orientation';
+import { OrientationBox } from '../base/orientation/OrientationBox';
+import { OrientationText } from '../base/orientation/OrientationText';
 import { Rect } from '../base/Rect';
 import { round } from '../base/utils/Math';
 import { MaxMeasure } from '../base/utils/MaxMeasure';
