+++ src/charts/VarianceChart.ts
@@ -4,8 +4,9 @@ import { Box } from '../base/Box';
 import { Color } from '../base/Color';
 import { Scenario, extractSimpleValue } from '../base/Data';
 import { SimpleDataType } from '../base/DataTypes';
-import { Orientation, OrientationBox } from '../base/OrientationBox';
-import { OrientationText } from '../base/OrientationText';
+import { Orientation } from '../base/orientation/Orientation';
+import { OrientationBox } from '../base/orientation/OrientationBox';
+import { OrientationText } from '../base/orientation/OrientationText';
 import { Rect } from '../base/Rect';
 import { MaxMeasure } from '../base/utils/MaxMeasure';
 import { Chart, ChartProps } from './Chart';
